Generalize file paths

+++ simulator.py
@@ -1,21 +1,30 @@
-import json, math, warnings, os,statistics, datetime as dt
+import math, statistics, datetime as dt
+import json, warnings, os
 import numpy as np
 import pandas as pd
 import matplotlib.pyplot as plt
+
+import git, sys
+git_root= git.Repo(os.path.abspath(''),
+                   search_parent_directories=True).git.rev_parse('--show-toplevel')
+sys.path.append(git_root)
 from models import returnGenerator, annuity, model
 from data import constants as const
 
 # For reference, something that has a 3% growth is a 0.03 return/rate and 1.03 yield. That's how I'll define return/rate and yield here
 
 DEBUG_LVL = 1 # LVL 1: Print success rate, save worst failure, show plot | LVL 2: Investigate each result 1 by 1
-SAVE_DIR = 'diagnostics/saved' 
+SAVE_DIR = os.path.join(git_root,'diagnostics/saved')
 TODAY = dt.date.today()
 TODAY_QUARTER = (TODAY.month-1)//3
 TODAY_YR = TODAY.year
 TODAY_YR_QT = TODAY_YR+TODAY_QUARTER*.25
 MONTE_CARLO_RUNS = 500 # takes 20 seconds to generate 5000
-for file in os.scandir(SAVE_DIR): # delete previously saved files
-    os.remove(file.path)
+if os.path.exists(SAVE_DIR):
+    for file in os.scandir(SAVE_DIR): # delete previously saved files
+        os.remove(file.path)
+else:
+    os.makedirs(SAVE_DIR)
 
 class Simulator:
     def __init__(self,param_vals,override_dict={}):
@@ -434,7 +443,7 @@ class Simulator:
             Values for each quarter.
 
         """
-        FLAT_INFLATION= options['FLAT_INFLATION']
+        FLAT_INFLATION= options['flat_inflation']
         time_ls= options['time_ls']
         working_qts= options['working_qts']
         FI_qts= options['FI_qts']
@@ -496,7 +505,9 @@ class Simulator:
 params = model.load_params()
 param_vals = {key:obj["val"] for (key,obj) in params.items()}
 
+simulator = Simulator(param_vals)
+simulator.main()
 
-if __name__ == '__main__':
-    simulator = Simulator(param_vals)
-    simulator.main()
+# if __name__ == '__main__':
+#     simulator = Simulator(param_vals)
+#     simulator.main()
